refactor: Moved interfaces an abstractions from implementation folders to abstractions folder in core module.
