+++ src/app/core/abstractions/entity-parser.interface.ts
