+++ src/app/core/services/http-rundown.service.ts
@@ -4,7 +4,7 @@ import {catchError, map, Observable} from 'rxjs'
 import {RundownService} from '../abstractions/rundown-service.interface';
 import {HttpErrorService} from './http-error.service';
 import {Rundown} from '../models/rundown';
-import { EntityParser } from './entity-parser.interface'
+import { EntityParser } from '../abstractions/entity-parser.interface'
 
 const RUNDOWN_URL: string = 'http://localhost:3005/api/rundowns'
